pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ems"
version = "0.1.0"
authors = [
  {name = "Your Name", email = "your.email@example.com"},
]
description = "Event Management System with collaborative editing features"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.5",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email_validator>=2.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/ems"
"Bug Tracker" = "https://github.com/yourusername/ems/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]